pyproject.toml
[project]
name = "dacboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyserial>=3.5",
    "pyvisa>=1.15.0",
    "qcodes>=0.52.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
